requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
et-xmlfile==2.0.0
fastapi==0.115.14
fastjsonschema==2.21.1
ffmpy==0.6.0
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
gradio==5.35.0
gradio-client==1.10.4
greenlet==3.2.3
groovy==0.1.2
htmlmin==0.1.12
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
llama-cpp-python==0.3.9
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mlxtend==0.23.4
narwhals==1.44.0
numpy==2.3.1
openpyxl==3.1.5
orjson==3.10.18
pandas==2.3.0
patsy==1.0.1
pillow==11.2.1
plotly==6.2.0
puremagic==1.29
pydub==0.25.1
pyparsing==3.2.3
python-multipart==0.0.20
pytz==2025.2
pywin32==310
pyyaml==6.0.2
requests==2.32.4
rich==14.0.0
ruff==0.12.1
safehttpx==0.1.6
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
statsmodels==0.14.4
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
